renames act-recursively to reduce-recursively
