+++ libs/core/util/src/index.ts
@@ -1,4 +1,3 @@
-export * from './lib/act-recursively';
 export * from './lib/calculate-tally-change';
 export * from './lib/combo-analyzer';
 export * from './lib/create-array';
@@ -11,6 +10,7 @@ export * from './lib/includes';
 export * from './lib/pluralize';
 export * from './lib/quiz';
 export * from './lib/record-to-array';
+export * from './lib/reduce-recursively';
 export * from './lib/select-random';
 export * from './lib/snake-case';
 export * from './lib/sort';
