+++ libs/core/util/src/lib/reduce-recursively.spec.ts
@@ -1,4 +1,4 @@
-import { actRecursively } from './act-recursively';
+import { actRecursively } from './reduce-recursively';
 
 interface Item {
   name: string;
