+++ libs/core/util/src/lib/reduce-recursively.ts
